Refactoring Pathway Age Search Page
