+++ templates/met_explore/pathway_age_search.html
@@ -51,8 +51,8 @@
       </p>
     </div>
 
-    <form action="pathway_search" method="get" class="form-inline my-2 my-lg-0">
-      <input id = "pathway_search" name ='pathway_search' class="form-control mr-sm-2" type="text" placeholder="Pathway Search">
+    <form action="pathway_age_search" method="get" class="form-inline my-2 my-lg-0">
+      <input id = "pathway_age_search" name ='pathway_age_search' class="form-control mr-sm-2" type="text" placeholder="Pathway Search">
       <button class="btn btn-outline-primary my-2 my-sm-0" type="submit">Search</button>
     </form>
 
@@ -84,7 +84,6 @@
                           <th class="col vw150">Age</th>
                           <th id = 'AF-pvalue' class="col vw150">Adult F/W p-value</th>
                           <th id = 'AM-pvalue'class="col vw150">Adult M/W p-value</th>
-                          <th id = 'L-pvalue'class="col vw150">Larvae/W p-value </th>
                         </tr>
                       </thead>
                       <tbody>
